Rename source_system provenance columns to short TINYINT form; merge Tier1 groups 1&4
created_source_system_id/updated_source_system_id -> created_src_id/updated_src_id,
INT -> TINYINT (V197) per Dharani's request: shorter names, and the FK never needs
more than 255 distinct source_system rows. Cascaded into rebuilding
source_system.source_system_id itself (FK type must match PK type exactly), and into
Java field types (Integer -> Short, required for Hibernate's strict TINYINT validation).

Also merges the 28 Tier1 entities completed by parallel worktree agents (groups 1 & 4
of the 8-way Tier1 rollout) before they hit a session rate limit, updated to the new
column naming. Live-verified end-to-end: Tier1 (JPA) and Tier2 (generic) creates/updates
both stamp provenance correctly against the real dev DB.

+++ etrm-system/etrm-backend/src/main/java/com/etrm/system/common/SourceSystemDefaults.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 
 /**
  * Resolves dbo.source_system ids for JPA entity provenance stamping
- * (created_source_system_id/updated_source_system_id, V194/V195) — a plain
+ * (created_src_id/updated_src_id, V194/V195) — a plain
  * {@code @MappedSuperclass} lifecycle callback (see AuditableEntity) can't
  * receive Spring-injected beans directly, so this component caches the
  * resolved id in a static field on startup instead, the standard escape
@@ -24,10 +24,15 @@ public class SourceSystemDefaults {
                 Integer.class, "TIER1_APPLICATION_SCREEN");
     }
 
-    public static int tier1ApplicationScreen() {
+    /** Returns {@code short} not {@code int} — dbo.source_system.source_system_id
+     *  is TINYINT (V197), and entity fields mapping it must be Short for
+     *  Hibernate's strict ddl-auto:validate type check to pass (found this
+     *  the hard way: an Integer-typed field fails validation against a
+     *  TINYINT column even though the value always fits). */
+    public static short tier1ApplicationScreen() {
         if (tier1ApplicationScreenId < 0) {
             throw new IllegalStateException("SourceSystemDefaults not yet initialized — accessed before Spring context startup completed.");
         }
-        return tier1ApplicationScreenId;
+        return (short) tier1ApplicationScreenId;
     }
 }
